Refactor error handling into SRP modules
